requirements.txt
albumentations>=1.4,<2.0
fastapi>=0.116,<1.0
matplotlib>=3.9,<4.0
numpy>=1.26,<3.0
opencv-python>=4.10,<5.0
pandas>=2.2,<3.0
pillow>=10.4,<12.0
pydantic>=2.8,<3.0
python-multipart>=0.0.9,<1.0
scikit-learn>=1.5,<2.0
seaborn>=0.13,<1.0
tensorflow>=2.16,<2.18
uvicorn>=0.30,<1.0
